fix failing tests
Some code implicitly assumed that the objects stored in a NodeList will always be moved when the array grows, e.g. when holding on to a pointer (or in this specific case, an iterator) into some member array/list. The OGDF Array always used move, while std::vector falls back to copying if the move constructor is (implicitly) noexcept(false) to ensure that the grow operation remain atomic. We should add a static check to ensure that all move operations are actually noexcept (RSPEC-5018).