SimpleCCPacker: Ensure edge bends are preserved after drawings
When using a layout algorithm that produces edge bends with the simple CC
packer layout, those were not available in the GraphAttributes passed as
parameter to SimpleCCPacker::call method.

The simple CC packer layout draws each connected component of a graph with
a provided layout algorithm then packs the resulting drawings.
However, the drawings of components with edge bends were not taken into account
in the previous implementation.

This commit ensures edge bends are preserved after each connected component
drawing but also taken into consideration when performing the packing operation.